pyproject.toml
[tool.poetry]
name = "autocam"
version = '1.1.3'
description = "Autocam"
authors = ["Petrus H Zwart <phzwart@lbl.gov>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/phzwart/autocam"
repository = "https://github.com/phzwart/autocam"
documentation = "https://autocam.readthedocs.io"
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/phzwart/autocam/releases"

[tool.poetry.dependencies]
python = "^3.9"
click = ">=8.0.1"
wheel = ">=0.38.1"
pydantic = "^2.5.0"
pyyaml = "^6.0"
jinja2 = "^3.0.0"

[tool.poetry.group.dev.dependencies]
Pygments = ">=2.10.0"
bandit = ">=1.7.4"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = '1.1.3'}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
setuptools-scm = ">=8.0.0"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = '1.1.3'}
myst-parser = {version = '1.1.3'}

[tool.poetry.scripts]
autocam = "autocam.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["autocam", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 85

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.black]
line-length = 88
target-version = ['py39']
exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | _version\.py
)/
'''

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true

[tool.setuptools_scm]
write_to = "src/autocam/_version.py"
version_file = "src/autocam/_version.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
